[Encoding] Teach specialize encodings to handle pad encodings. (#20845)
This allows the pad based encodings that are not serializable to be
specialized to serialized versions during `SpecializeEncodings` pass.
This requires handling `?` entries in the `pad_encoding_layout`.

This PR changes the `getLayout` method of `GPUPadLayoutAttr` to not
handle `matmul_k` encoding attribute, some of the tests in
`materialize_encoding_into_padding.mlir` now fail. Those are removed
since they will be adapted/folded into `materialize_encoding.mlir` in a
follow up.

Towards #20835

---------

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>
Signed-off-by: Jorn Tuyls <jorn.tuyls@gmail.com>
Co-authored-by: Jorn Tuyls <jorn.tuyls@gmail.com>