[NFC] Refresh the interface names for Encoding dialect and data-tiling specifics. (#20985)
The IREE team has been developing the encodings and data-tiling
specifics for a while. Now we have better pictures about the layering
and the naming. The revision updates the interface names, which removes
the redundant words and provides more descriptive names.

The attributes and interfaces are already defined within encoding
namespace/dialect. There is a hidden convention that we prefer fully
specify type name in IREE core dialects. Because it is easier to do
large scale replacements, avoids conflicts with MLIR upstream types, and
makes it easier to port passes between dialects. In this context,
Encoding is redundant in attribute names and interface names.

- `EncodingLayoutResolverAttrInterface` -> `LayoutResolverAttr`.
- `SerializableEncodingAttrInterface` -> `SerializableAttr`.
- `LayoutAttrInterface` -> `LayoutMaterializerAttr`.
- `PackedLayoutAttrInterface` -> `PackedLayoutMaterializerAttr `.

The Host/Device words are removed from the interface implementation
because they were introduced for having a clear line between host and
device. We have better abstraction now, and they are no longer needed.

It is a step towards https://github.com/iree-org/iree/issues/20742

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>