Adding `iree_hal_amdgpu_executable_t` implementation + no-op cache. (#21040)
The executable is verified and loaded once and then uploaded to all
agents in the topology. Future work can allow for heterogeneous GPU
agents but today all are assumed to be the same.

A copy of the per-export `iree_hal_amdgpu_device_kernel_args_t` metadata
is retained on the host as well as stored locally on each agent. This
allows the host to verify dispatch information during command buffer
recording and the devices to avoid hitting the bus to resolve kernel
parameters. Though we may want to optimize dispatch overhead by
pre-baking kernel metadata into command buffers we still want the
device-side copies for dynamic dispatch where we decide which kernel to
dispatch immediately prior to doing so on the device.