[GPU] Handle transient private values in control-flow when prefetching (#21037)
This PR makes the stage selection for read restricted by making sure
only global reads are chosen as root read ops. This is because private
reads dont necessarily belong to read stage and the valueDependency
function should be able to put those in the correct stage.
Additionally the PR adds support for control flow with private writes
that should be placed in write stage under the appropriate conditions.

Signed-off-by: Nirvedh Meshram <nirvedh@gmail.com>