[Dispatch] Only bubble reshapes when possibly blocking fusion (#20108)
This change prevents reshape propagation when the reshape is known to
not be blocking fusion. For example, when a producer/consumer is a
`flow.tensor.bitcast` or the reshape is on the edge of the program. This
has 2 benefits: (1) prevents unit dims from getting propagated since
they don't get dropped from flow/hal/etc ops, leading to reshapes next
to these ops to expand/collapse unit dims. (2) reshape propagation is
imperfect so stopping early when the reshape definitely won't block
fusion is better then propagating it and getting stuck.


Prevents propagating the unit dims causing
https://github.com/iree-org/iree/issues/19991

---------

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>