[Codegen][GPU] Sort intrinsic according to k alignment - Step 1 of 2- Track MmaInterfaceAttr via field instead of index (#21103)
This PR is NFC. This PR simplified the intrinsic derive process via:
- Created a dedicated `GPUIntrinsicType` to differentiate from a
`GPUMatmulShapeType`
- This newly created `GPUIntrinsicType` inherits from
`GPUMatmulShapeType`
   - It keep track of the `MmaInterfaceAttr` directly in a field
- Removed the need to track the intrinsic separately through a index
like `intrinsicIndex`
- The separately maintained index used reference only to the position of
intrinsic, making it error prone. To sort and prioritize the intrinsic
will have to keep track of the intrinsic before and after sorting.
- Keeping track of intrinsic instead of index makes tracking much easier
- Replaced the index field from `GPUMMASchedule` to `MmaInterfaceAttr`
for the same reasoning

Sort intrinsic according to K alignment is dependent on this change to
land. I break the NFC aspect of the feature into this PR so it is easier
to do code reviews. For the convolution configurations that are not
k-aligned by 16x16x16 intrinsic (but aligned with 32x32x8), this
together with next PR will bring 5-10x perf lift.

---------

Signed-off-by: jerryyin <zhuoryin@amd.com>