[compiler] remove uses of memref::ExpandOps pass (#21113)
This pass currently performs two rewrites:

1. *incorrectly* rewrite `atomic_rmw` for floating-point min/max, see
#21101;
  2. rewrite `reshape` into `reinterpret_cast`.

The rewrite 1 was added years ago, prior to LLVM IR supporting
floating-point atomics sufficiently and is no longer necessary. I will
remove it upstream anyway.

The rewrite 2 looks like a canonicalizaiton and it is unclear why it was
put into a separate pass, it will likely not remain there. It may have
some minor effect on performance if this rewrite is important to
simplify address computation, at which point it belongs to the set of
canonicalization patterns.

In general, I highly suspect that the usage of this pass was either
cargo-culted from existing sub-pipelines or confused with
`--expand-strided-metadata`, which is both necessary for lowering and
beneficial for address computation performance.

Signed-off-by: Alex Zinenko <git@ozinenko.com>