[LinalgExt] Add simple vectorization for map_scatter (#21090)
Allows vector types for the map_scatter input operand, and adds a
trivial vectorization pass, which converts the input of map_scatter ops
to vector operands with a vector.transfer_read.

This vectorization lowering is not complete, because the map_scatter op
in this form will not be able to lower into other operations. The final
lowering will happen as a decomposition after the map_scatter op has
been both vectorized and bufferized, and the map_scatter op will be
decomposed into a vector.scatter op. The reason this decomposition will
need to happen after bufferization, is that the vector.scatter op
requires a memref base. The decomposition pattern of the vectorized
map_scatter will come as a follow-up PR.

Eventually, the vectorized map_scatter op could instead be decomposed
into an `iree_vector_ext.transfer_scatter` op, but since the op does not
exist yet, the current plan is to use `vector.scatter` first.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>