[DT] Improve encoding hoisting pass. (#21275)
The revision runs ConstExprHoistingPolicy and hoist the chain ops
outside dispatch if they are const-exprs. Furthermore, it allows
hoisting tensor.extract_slice ops when they are mappable into Flow ops.
It also allows constant-like ops because the ConstExpr analysis has an
assumption that any root op is already hoisted to initializers. The rest
of operations are not allowed to be hoisted out atm.

The hoistEncodingsForConstExpr is dropped because it is no longer
needed. It was created for padding strategy.

The pass is changed to ModuleOp scope because it requires the analysis
for const-exprs. Otherwise, it causes a thread race condition.

Fixes https://github.com/iree-org/iree/issues/21100

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>