[DispatchCreation] Add a pass to split long running reduction loops. (#21280)
For long running reductions it is better to split the reduction into two
dispatches, the first dispatch running partial reductions and the second
dispatch combining the result of these partial reductions.

The process of splitting the reduction is essentially a tiling
transformation, which is supported upstream using `scf::tileUsingSCF`
with the `ReductionTilingStrategy::PartialReductionOuterParallel`
strategy. This PR adds a pass `FormSplitReductionDispatchesPass` that
uses this strategy in IREE.

See more details : https://hackmd.io/-iRV4xydSIKzsup9fJ5xbg

The tiling method creates a `scf.forall` loop that runs the partial
reduction in parallel. These are moved into their own dispatches. The
combination of these partial reductions (which are linalg operations)
are left outside and get moved into their own dispatches in subsequent
passes.

Couple of other changes are needed

- An attribute `SplitReductionMappingAttr` is added that implements the
`DeviceMappingAttrInterface` that can be used as a mapping attribute
with `scf.forall` created for the parallel partial reductions.

- The operation within the `scf.forall` is meant to be further
distributed on GPUs (just as if it were not within the `scf.forall`).
The number of workgroups needed to execute the parallel partial
reductions is the number of workgroups needed to execute the op by
itself, multiplied by the number of partial reductions. This is
represented by a new op
`iree_tensor_ext.dispatch_workgroup_count_split_reduction_modifier`. The
exact computation of the number of workgroups is done by the backend.

Signed-off-by: MaheshRavishankar <mravisha@amd.com>
Co-authored-by: Ian Wood <ianwood2024@u.northwest.edu>